refactor: Restructure the _poems routes and related code
