+++ app/routes/_poems/route.tsx
@@ -1,0 +1,9 @@
+import { Outlet } from "@remix-run/react";
+
+export default function PoemsLayout() {
+  return (
+    <div className="relative max-w-[1920px]">
+      <Outlet />
+    </div>
+  );
+}
